# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TinyURL API Document",
        "title": "TinyURL Swagger API Docs",
        "contact": {
            "name": "AdolphLWQ",
            "url": "https://git.io/tinyurl",
            "email": "kenan3015@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/shorten": {
            "post": {
                "description": "shortening origin url and return short path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "shortening origin url",
                "parameters": [
                    {
                        "description": "origin url",
                        "name": "shortpath",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PostURL"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "short path",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "consumes": [
                    "application/x-json-stream"
                ],
                "summary": "健康检查",
                "responses": {
                    "200": {
                        "description": "health",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/n/{shortpath}": {
            "get": {
                "description": "get source url by short path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "parse shorten path and return source url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short path",
                        "name": "shortpath",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "304": {
                        "description": "haha",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.PostURL": {
            "type": "object",
            "properties": {
                "oritin_url": {
                    "type": "string"
                }
            }
        }
    }
}